discretization: add an explicit cast to avoid sign extension
Fixes an error reported by Coverity: "sign_extension: Suspicious
implicit sign extension: this->m_dimensions with type uint8_t
(8 bits, unsigned) is promoted in this->m_dimensions * nCoeffs to
type int (32 bits, signed), then sign-extended to type unsigned
long (64 bits, unsigned). If this->m_dimensions * nCoeffs is
greater than 0x7FFFFFFF, the upper bits of the result will all
be 1".

+++ src/discretization/reconstruction.cpp
@@ -1612,7 +1612,7 @@ void ReconstructionPolynomial::computeGradients(int degree, const std::array<dou
     // Generic polynomial
     int nCoeffs = ReconstructionPolynomial::getCoefficientCount(degree, m_dimensions);
 
-    BITPIT_CREATE_WORKSPACE(dcsi, double, m_dimensions * nCoeffs, 3 * MAX_STACK_WORKSPACE_SIZE);
+    BITPIT_CREATE_WORKSPACE(dcsi, double, static_cast<std::size_t>(m_dimensions * nCoeffs), 3 * MAX_STACK_WORKSPACE_SIZE);
     ReconstructionPolynomial::evalPointBasisDerivatives(degree, m_dimensions, m_origin, point, {{1., 0., 0.}}, dcsi);
     ReconstructionPolynomial::evalPointBasisDerivatives(degree, m_dimensions, m_origin, point, {{0., 1., 0.}}, dcsi + nCoeffs);
     if (m_dimensions == 3) {
@@ -1846,7 +1846,7 @@ void ReconstructionPolynomial::computeGradientsLimited(int degree, const std::ar
     int nCoeffs = ReconstructionPolynomial::getCoefficientCount(degree, m_dimensions);
     const std::vector<uint16_t> &nDegreeCoeffs = ReconstructionPolynomial::getDegreeCoefficientsCount(m_dimensions);
 
-    BITPIT_CREATE_WORKSPACE(dcsi, double, m_dimensions * nCoeffs, 3 * MAX_STACK_WORKSPACE_SIZE);
+    BITPIT_CREATE_WORKSPACE(dcsi, double, static_cast<std::size_t>(m_dimensions * nCoeffs), 3 * MAX_STACK_WORKSPACE_SIZE);
     ReconstructionPolynomial::evalPointBasisDerivatives(degree, m_dimensions, m_origin, point, {{1., 0., 0.}}, dcsi);
     ReconstructionPolynomial::evalPointBasisDerivatives(degree, m_dimensions, m_origin, point, {{0., 1., 0.}}, dcsi + nCoeffs);
     if (m_dimensions == 3) {
